requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
python-dotenv
redis
rq
python-dateutil
holidays
streamlit
requests
